Overhauled Maine/Nebraska functionality
Changed Maine/Nebraska votes from form inputs to individual buttons to match the respective district/popular electoral votes. These function the same as the states so are more user-friendly. Added a similar button for District of Columbia since the location was small and difficult to click. Adjusted some responsive styling.